Move registry module to crate level.
