+++ src/internal/mod.rs
@@ -1,4 +1,3 @@
 pub(crate) mod query;
-pub(crate) mod registry;
 #[cfg(feature = "parallel")]
 pub(crate) mod system;
