--- src/internal/registry/mod.rs
@@ -1,30 +1,0 @@
-mod debug;
-mod eq;
-mod length;
-mod send;
-#[cfg(feature = "serde")]
-mod serde;
-mod storage;
-mod sync;
-
-#[cfg(feature = "serde")]
-pub(crate) use self::serde::{RegistryDeserialize, RegistrySerialize};
-pub(crate) use debug::RegistryDebug;
-pub(crate) use eq::{RegistryEq, RegistryPartialEq};
-pub(crate) use send::RegistrySend;
-pub(crate) use sync::RegistrySync;
-
-use crate::{component::Component, registry::Null};
-use length::RegistryLength;
-use storage::RegistryStorage;
-
-pub trait RegistrySeal: RegistryLength + RegistryStorage {}
-
-impl RegistrySeal for Null {}
-
-impl<C, R> RegistrySeal for (C, R)
-where
-    C: Component,
-    R: RegistrySeal,
-{
-}
