+++ src/lib.rs
@@ -5,6 +5,7 @@ extern crate alloc;
 pub mod component;
 pub mod entities;
 pub mod entity;
+pub mod registry;
 pub mod world;
 
 #[doc(hidden)]
