+++ src/public/mod.rs
@@ -1,3 +1,2 @@
 pub mod query;
-pub mod registry;
 pub mod system;
