+++ src/registry/debug.rs
