+++ src/registry/eq.rs
