+++ src/registry/seal/length.rs
@@ -1,17 +1,17 @@
 use crate::{component::Component, registry::Null};
 
-pub trait RegistryLength {
+pub trait Length {
     const LEN: usize;
 }
 
-impl RegistryLength for Null {
+impl Length for Null {
     const LEN: usize = 0;
 }
 
-impl<C, R> RegistryLength for (C, R)
+impl<C, R> Length for (C, R)
 where
     C: Component,
-    R: RegistryLength,
+    R: Length,
 {
     const LEN: usize = R::LEN + 1;
 }
