+++ src/world/impl_eq.rs
@@ -1,5 +1,5 @@
 use super::World;
-use crate::internal::registry::{RegistryEq, RegistryPartialEq};
+use crate::registry::{RegistryEq, RegistryPartialEq};
 
 impl<R> PartialEq for World<R>
 where
