+++ src/world/impl_send.rs
@@ -1,3 +1,3 @@
-use crate::{internal::registry::RegistrySend, world::World};
+use crate::{registry::RegistrySend, world::World};
 
 unsafe impl<R> Send for World<R> where R: RegistrySend {}
