+++ src/world/impl_sync.rs
@@ -1,3 +1,3 @@
-use crate::{internal::registry::RegistrySync, world::World};
+use crate::{registry::RegistrySync, world::World};
 
 unsafe impl<R> Sync for World<R> where R: RegistrySync {}
